Move key_wrap and verifier to public headers
