+++ src/apps/js_v8/tmpl/ccf_global.cpp
@@ -3,12 +3,13 @@
 #include "ccf_global.h"
 
 #include "apps/js_v8/tmpl/crypto.h"
+#include "ccf/crypto/entropy.h"
+#include "ccf/crypto/key_wrap.h"
 #include "ccf/crypto/rsa_key_pair.h"
 #include "ccf/crypto/sha256.h"
+#include "ccf/crypto/verifier.h"
 #include "ccf/ds/logger.h"
 #include "consensus.h"
-#include "crypto/entropy.h"
-#include "crypto/key_wrap.h"
 #include "historical.h"
 #include "historical_state.h"
 #include "kv_store.h"
