+++ src/apps/js_v8/tmpl/crypto.cpp
@@ -3,9 +3,9 @@
 #include "apps/js_v8/tmpl/crypto.h"
 
 #include "ccf/crypto/rsa_key_pair.h"
+#include "ccf/crypto/verifier.h"
 #include "ccf/ds/logger.h"
 #include "crypto/ecdsa.h"
-#include "crypto/verifier.h"
 #include "template.h"
 
 // NOTE: The rest of the crypto functions are defined in ccf_global.cpp.
