+++ src/crypto/key_wrap.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 
-#include "key_wrap.h"
+#include "ccf/crypto/key_wrap.h"
 
 #include "ccf/crypto/key_pair.h"
 #include "ccf/crypto/rsa_key_pair.h"
