+++ src/crypto/openssl/verifier.h
@@ -2,8 +2,8 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
+#include "ccf/crypto/verifier.h"
 #include "crypto/openssl/openssl_wrappers.h"
-#include "crypto/verifier.h"
 
 #include <openssl/x509.h>
