+++ src/endpoints/authentication/sig_auth.cpp
@@ -3,6 +3,7 @@
 
 #include "ccf/endpoints/authentication/sig_auth.h"
 
+#include "ccf/crypto/verifier.h"
 #include "ds/lru.h"
 #include "enclave/rpc_context.h"
 #include "http/http_sig.h"
