+++ src/node/history.h
@@ -3,7 +3,6 @@
 #pragma once
 
 #include "ccf/ds/logger.h"
-#include "crypto/verifier.h"
 #include "ds/dl_list.h"
 #include "ds/thread_messaging.h"
 #include "endian.h"
