+++ src/tls/test/main.cpp
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #include "ccf/crypto/key_pair.h"
+#include "ccf/crypto/verifier.h"
 #include "ccf/ds/buffer.h"
 #include "ccf/ds/logger.h"
 #include "crypto/certs.h"
-#include "crypto/verifier.h"
 #include "tls/client.h"
 #include "tls/msg_types.h"
 #include "tls/server.h"
